@@ -11,7 +11,8 @@ authors = [
 ]
 readme = "README.md"
 license = {text = "MIT"}
-requires-python = ">=3.9"
+# Empfohlene Python-Version: 3.11+
+requires-python = ">=3.11"
 keywords = ["ai", "media", "analysis", "video", "audio", "image", "processing"]
 classifiers = [
     "Development Status :: 4 - Beta",
