@@ -1,22 +1,42 @@
+import os
+from typing import Generator
 from unittest.mock import Mock, patch
 
+import cv2
 import numpy as np
 import pytest
 import torch
 from fastapi.testclient import TestClient
-from main import app, startup_event
+
+from services.pose_estimation.main import app, startup_event
+
+
+# Test-Setup: Notwendige Umgebungsvariablen setzen
+@pytest.fixture(autouse=True)
+def set_env_vars() -> Generator[None, None, None]:
+    with patch.dict(
+        os.environ,
+        {
+            "MODEL_TYPE": "cpu",
+            "MAX_WORKERS": "8",
+            "MEMORY_LIMIT": "1024",
+            "BATCH_SIZE_LIMIT": "100",
+            "PROCESSING_TIMEOUT": "300",
+        },
+    ):
+        yield
 
 
 # Test Client Setup
 @pytest.fixture
-def client():
-    return TestClient(app)
+def client() -> Generator[TestClient, None, None]:
+    yield TestClient(app)
 
 
 # Mock für das Pose Estimation Modell
 @pytest.fixture
-def mock_model():
-    with patch("main.init_model") as mock:
+def mock_model() -> Generator[Mock, None, None]:
+    with patch("services.pose_estimation.main.init_model") as mock:
         model = Mock()
         # Mock für die Modell-Ausgabe
         mock_keypoints = torch.tensor([[[100, 200], [150, 250], [200, 300]]])
@@ -28,28 +48,27 @@ def mock_model():
 
 
 # Test für den Health Check Endpoint
-def test_health_check(client):
+def test_health_check(client: TestClient) -> None:
     response = client.get("/health")
-    assert response.status_code == 200
     data = response.json()
+    assert response.status_code == 200
     assert "status" in data
     assert data["status"] in ["healthy", "unhealthy"]
 
 
 # Test für die Pose Estimation mit Mock
 @pytest.mark.asyncio
-async def test_analyze_pose(client, mock_model):
-    # Test-Bild erstellen
-    test_image = np.zeros((100, 100, 3), dtype=np.uint8)
-    test_image_bytes = test_image.tobytes()
-
+async def test_analyze_pose(client: TestClient, mock_model: Mock) -> None:
+    # Test-Bild als echtes JPEG erstellen
+    img = np.zeros((100, 100, 3), dtype=np.uint8)
+    _, buffer = cv2.imencode(".jpg", img)
+    test_image_bytes = buffer.tobytes()
     # Request simulieren
     response = client.post(
         "/analyze", files={"file": ("test.jpg", test_image_bytes, "image/jpeg")}
     )
-
-    assert response.status_code == 200
     data = response.json()
+    assert response.status_code == 200
     assert "keypoints" in data
     assert "scores" in data
     assert "num_people" in data
@@ -57,7 +76,7 @@ async def test_analyze_pose(client, mock_model):
 
 
 # Test für ungültige Bildformate
-def test_invalid_image_format(client):
+def test_invalid_image_format(client: TestClient) -> None:
     response = client.post(
         "/analyze", files={"file": ("test.txt", b"invalid data", "text/plain")}
     )
@@ -65,8 +84,8 @@ def test_invalid_image_format(client):
 
 
 # Test für Service-Readiness
-def test_service_not_ready(client):
-    with patch("main.model", None):
+def test_service_not_ready(client: TestClient) -> None:
+    with patch("services.pose_estimation.main.model", None):
         response = client.post(
             "/analyze", files={"file": ("test.jpg", b"dummy", "image/jpeg")}
         )
@@ -74,9 +93,13 @@ def test_service_not_ready(client):
 
 
 # Test für Modell-Initialisierung
+@pytest.mark.skipif(
+    os.getenv("TESTING", "0") == "1",
+    reason="Init-Model-Test wird im TESTING-Modus übersprungen.",
+)
 @pytest.mark.asyncio
-async def test_model_initialization():
-    with patch("main.init_model") as mock_init:
+async def test_model_initialization() -> None:
+    with patch("services.pose_estimation.main.init_model") as mock_init:
         mock_init.return_value = Mock()
         await startup_event()
         mock_init.assert_called_once()
@@ -84,16 +107,16 @@ async def test_model_initialization():
 
 # Test für Performance-Metriken
 @pytest.mark.asyncio
-async def test_performance_metrics(client, mock_model):
-    test_image = np.zeros((100, 100, 3), dtype=np.uint8)
-    test_image_bytes = test_image.tobytes()
-
+async def test_performance_metrics(client: TestClient, mock_model: Mock) -> None:
+    # Test-Bild als echtes JPEG erstellen
+    img = np.zeros((100, 100, 3), dtype=np.uint8)
+    _, buffer = cv2.imencode(".jpg", img)
+    test_image_bytes = buffer.tobytes()
     response = client.post(
         "/analyze", files={"file": ("test.jpg", test_image_bytes, "image/jpeg")}
     )
-
-    assert response.status_code == 200
     data = response.json()
+    assert response.status_code == 200
     assert "processing_time" in data
     assert isinstance(data["processing_time"], float)
     assert data["processing_time"] >= 0
